refactor: move the dialog methods into extension methods
Instead of having a class for DialogHandler just to handle the two methods, move the two methods
into a file for extension methods. Since the class isn't really needed anyway this provides some
slight clean up
